pyproject.toml
[project]
name            = "caseus"
version         = "0.1.0"
description     = "An asynchronous Transformice networking library"
# TODO: readme  = "README.md"
requires-python = ">=3.7"
license         = {file = "LICENSE"}

authors = [
    {name = "friedkeenan"},
]

classifiers = [
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: GNU General Public License v2 (GPLv2)",
    "Operating System :: OS Independent",
]

dependencies = [
    "pak",

    "importlib_metadata; python_version < '3.8'",
    "atpublic",
    "fixedint",
    "aioconsole",
    "scapy",
]

[project.urls]
repository = "https://github.com/friedkeenan/caseus"
